+++ service.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"os/signal"
 	"path"
-	"runtime"
 	"strings"
 	"syscall"
 
@@ -326,17 +325,4 @@ func openLogFile(logger *log.Logger, logPath string) {
 
 	signal.Notify(sigs, syscall.SIGHUP)
 	<-logging
-}
-
-func dumpGoroutineStacks(logger *log.Logger) {
-	sigs := make(chan os.Signal, 1)
-	go func() {
-		for {
-			<-sigs
-			buf := make([]byte, 1<<16)
-			size := runtime.Stack(buf, true)
-			logger.Infof("*** goroutine dump ***\n%s", buf[:size])
-		}
-	}()
-	signal.Notify(sigs, syscall.SIGUSR1)
 }
